Honor Phab URLs in log messages.
If exactly one valid Phab URL is present in a log message, use that as
the review instead of matching on the title.  For 'git arc list' this
means this can find commits not present in 'arc list', so a similar
status line is constructed (but without colors) from the data returned
by the phid.lookup conduit call.